Move custom backend related code in main.ts to its own file
