+++ electron/src/main.ts
@@ -30,9 +30,6 @@ import {
   OnHeadersReceivedResponse as HeadersReceivedResponse,
   shell,
   WebContents,
-  webContents,
-  session,
-  dialog,
 } from 'electron';
 import * as fs from 'fs-extra';
 import {getProxySettings} from 'get-proxy-settings';
@@ -68,12 +65,10 @@ import {ProxyPromptWindow} from './window/ProxyPromptWindow';
 import {WindowManager} from './window/WindowManager';
 import {WindowUtil} from './window/WindowUtil';
 import ProxyAuth from './auth/ProxyAuth';
-import {WindowUrl} from './window/WindowUrl';
 import WindowStateKeeper = require('electron-window-state');
 import fileUrl = require('file-url');
 import {changeEnvironmentPrompt} from './lib/changeEnvironmentPrompt';
-import * as parseCsp from 'content-security-policy-parser';
-import * as buildCsp from 'content-security-policy-builder';
+import {injectCustomBackend} from './lib/CustomBackend';
 
 const APP_PATH = path.join(app.getAppPath(), config.electronDirectory);
 const INDEX_HTML = path.join(APP_PATH, 'renderer/index.html');
@@ -505,15 +500,6 @@ const allowCORSForLocalhost = (contents: WebContents) => {
   contents.session.webRequest.onHeadersReceived(filter, listenerOnHeadersReceived);
 };
 
-const addCustomBackendToCSP = (originalCsp: string, backendUrl: string, directive: string = 'default-src') => {
-  const csp = parseCsp(originalCsp[0]);
-  // Note: The backend only may not be the only host we want to whitelist?
-  csp[directive].push(backendUrl);
-  return buildCsp({
-    directives: csp,
-  });
-};
-
 class ElectronWrapperInit {
   logger: logdown.Logger;
 
@@ -574,52 +560,7 @@ class ElectronWrapperInit {
             webPreferences.webSecurity = true;
             webPreferences.enableBlinkFeatures = '';
 
-            // Custom backend logic implementation
-            const src = new URL(params.src);
-            const customBackend = src.searchParams.get('backendUrl');
-            if (customBackend) {
-              this.logger.log('Using custom backend for a webview...');
-              // Remove backend url here to preserve privacy
-              src.searchParams.delete('backendUrl');
-              params.src = src.toString();
-
-              // Get the session of the partition
-              const currentSession =
-                params.partition && params.partition !== ''
-                  ? session.fromPartition(params.partition)
-                  : session.defaultSession;
-              if (currentSession) {
-                // Intercept requests
-                const currentWebappUrl = EnvironmentUtil.URL_WEBAPP[EnvironmentUtil.getEnvironment()];
-                const filter = {urls: [`${currentWebappUrl}/*`, `${customBackend}/*`]};
-                const listenerOnHeadersReceived = (
-                  details: OnHeadersReceivedListenerDetails,
-                  callback: (response: HeadersReceivedResponse) => void,
-                ) => {
-                  const responseHeaders = (details as any).responseHeaders;
-                  const extraHeaders: Record<string, string[]> = {};
-
-                  const hasCORS =
-                    responseHeaders['Access-Control-Allow-Credentials'] ||
-                    responseHeaders['Access-Control-Allow-Origin'];
-                  if (hasCORS) {
-                    extraHeaders['Access-Control-Allow-Credentials'] = ['true'];
-                    extraHeaders['Access-Control-Allow-Origin'] = [currentWebappUrl];
-                  }
-
-                  const currentCSP = responseHeaders['Content-Security-Policy'];
-                  if (currentCSP) {
-                    extraHeaders['Content-Security-Policy'] = [addCustomBackendToCSP(currentCSP, customBackend)];
-                  }
-
-                  callback({
-                    cancel: false,
-                    responseHeaders: {...details.responseHeaders, ...extraHeaders},
-                  });
-                };
-                currentSession.webRequest.onHeadersReceived(filter, listenerOnHeadersReceived);
-              }
-            }
+            injectCustomBackend(event, webPreferences, params, BASE_URL);
           });
           break;
         }
